backend/requirements.txt
torch
chromadb
sentence-transformers
google-generativeai
flask
flask-cors
torch